Add support for setting file attributes in vector stores
The official OpenAI API for .NET does not support this, so we resort to the lower-level HTTP API to achieve this.

In order to get parity with the rendering on the output for both listing and adding, we also switch to rendering raw responses. We should do the same for all the other APIs, since the documented JSON payloads will otherwise not match what we render.