File metadata, lists filtering and proper JSON output
We switch now to rendering the underlying JSON response from the APIs, so that they match the documentation from OpenAI for further JQ querying, which is far better than serializing the C#-view of the generated APIs.

This is particularly useful for cases where re-serializing loses data, such as the attributes associated with a vector file.

We fix setting file attributes for a vector store file, which is key to performing semantic file search downsteam.

Finally, we provide --skip, --take and more importantly --range (with C# range expression syntax) for slicing the results from the list commands, since otherwise they can be quite overwhelming.

For large lists, also took the chance to do live rendering while we're collecting results, and a final table rendering for whole-table output without loss of data.