Improve performance with --skip --take on large lists
Rather than reading all entries before applying the filters, we can be smarter about --skip and --take and exit the enumeration as soon as we have enough items.

Doing the same for Range is tricky since it can represent a lot of cases, so for now we keep full evaluation for range expressions.